requirements_cpu.txt
# torch==2.3.0+cpu --index-url https://download.pytorch.org/whl/cpu
# transformers>=4.41.0
# accelerate>=0.31.0
# datasets>=2.19.0
# evaluate>=0.4.2
# pyyaml>=6.0.1
# docker>=7.0.0
# gitpython>=3.1.43
# tqdm>=4.66.4
# psutil>=5.9.8
# numpy>=1.26.4
# pandas>=2.2.2

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.0+cpu
torchvision==0.18.0+cpu
torchaudio==2.3.0+cpu
transformers>=4.41.0
accelerate>=0.31.0
datasets>=2.19.0
evaluate>=0.4.2
pyyaml>=6.0.1
gitpython>=3.1.43
tqdm>=4.66.4
psutil>=5.9.8
numpy>=1.26.4
pandas>=2.2.2
